Rotary enc update
A few days' worth of work trying to get the rotary encoder working.
As of writing this the outputs from the encoder can be read, just not interpreted so well.
Having issues with bounce and debouncing code.
This update comes as part of a package deal with some readme updates too.