This was way harder than it needed to be.
I was copying the .git folder last time too. That probably messes a lot of stuff up. Explains why it thought it was a different branch than what it actually was.

This Would've just been the update for getting the rotary encoder to work with the ClickEncoder library and getting the potentiometer to work with a simple analogRead, but no, I goofed up with a massive file (the Visuino installer) that I'd dumped in the code backups folder and promptly forgot about.
Now I have to somehow merge this into master or make this the new master or something somehow.